test: 테스트코드 경로 수정
